Revert `subspace-runtime-primitives` changes and move code to `subspace-node-primitives` as it is only needed there
