+++ crates/subspace-node-primitives/src/lib.rs
@@ -17,10 +17,39 @@
 //! Node primitives for Subspace Network.
 
 use futures::Future;
+use parity_scale_codec::{Decode, Encode};
+use scale_info::TypeInfo;
+use serde::{Deserialize, Serialize};
+use sp_application_crypto::KeyTypeId;
+use sp_core::bytes;
+use sp_runtime::traits::Hash as HashT;
 use std::pin::Pin;
-use subspace_runtime_primitives::{
-    BlockNumber, CollatorPair, Hash, HeadData, PersistedValidationData,
-};
+use subspace_runtime_primitives::{BlockNumber, Hash};
+
+/// Parachain head data included in the chain.
+#[derive(
+    Debug,
+    Default,
+    PartialEq,
+    Eq,
+    Clone,
+    PartialOrd,
+    Ord,
+    Hash,
+    Serialize,
+    Deserialize,
+    Encode,
+    Decode,
+    TypeInfo,
+)]
+pub struct HeadData(#[serde(with = "bytes")] pub Vec<u8>);
+
+impl HeadData {
+    /// Returns the hash of this head data.
+    pub fn hash(&self) -> Hash {
+        sp_runtime::traits::BlakeTwo256::hash(&self.0)
+    }
+}
 
 pub struct Collation {
     pub number: BlockNumber,
@@ -46,6 +75,17 @@ pub type CollationSecondedSignal = Vec<u8>;
 // TODO: SubspaceBlockWeight
 /// The cumulative weight of a block in a fork-choice rule.
 pub type BlockWeight = u32;
+
+#[derive(Debug, Default, PartialEq, Eq, Clone, Encode, Decode, TypeInfo)]
+pub struct PersistedValidationData<H = Hash, N = BlockNumber> {
+    // TODO: use a proper wrapper type?
+    /// The encoded optional parent head hash.
+    pub parent_head: Vec<u8>,
+    /// The relay-chain block number this is in the context of.
+    pub relay_parent_number: N,
+    /// The relay-chain block storage root this is in the context of.
+    pub relay_parent_storage_root: H,
+}
 
 impl CollationResult {
     /// Convert into the inner values.
@@ -73,6 +113,25 @@ pub type CollatorFn = Box<
         + Send
         + Sync,
 >;
+
+/// The key type ID for a collator key.
+const COLLATOR_KEY_TYPE_ID: KeyTypeId = KeyTypeId(*b"coll");
+
+mod collator_app {
+    use super::COLLATOR_KEY_TYPE_ID;
+    use sp_application_crypto::{app_crypto, sr25519};
+
+    app_crypto!(sr25519, COLLATOR_KEY_TYPE_ID);
+}
+
+/// Identity that collators use.
+pub type CollatorId = collator_app::Public;
+
+/// A Parachain collator keypair.
+pub type CollatorPair = collator_app::Pair;
+
+/// Signature on candidate's block data by a collator.
+pub type CollatorSignature = collator_app::Signature;
 
 /// Configuration for the collation generator
 pub struct CollationGenerationConfig {
