+++ cumulus/client/cirrus-service/Cargo.toml
@@ -32,4 +32,4 @@ cirrus-client-executor = { path = "../cirrus-executor" }
 
 # Subspace
 subspace-node = { path = "../../../crates/subspace-node" }
-subspace-runtime-primitives = { path = "../../../crates/subspace-runtime-primitives" }
+subspace-node-primitives = { path = "../../../crates/subspace-node-primitives" }
