+++ polkadot/node/collation-generation/src/lib.rs
@@ -32,8 +32,8 @@ use polkadot_node_subsystem_util::{
 };
 use std::sync::Arc;
 
-use subspace_node_primitives::CollationGenerationConfig;
-use subspace_runtime_primitives::{Hash, PersistedValidationData};
+use subspace_node_primitives::{CollationGenerationConfig, PersistedValidationData};
+use subspace_runtime_primitives::{Hash};
 
 mod error;
